# Implement a faster encoding algoritm for varint