--- a/AUTHORS
+++ b/AUTHORS
@@ -1,2 +1,3 @@
 opensource@google.com
 bburnett@centeredgesoftware.com
+info@couchbase.com