--- a/Snappier.Benchmarks/VarIntEncodingWrite.cs
+++ b/Snappier.Benchmarks/VarIntEncodingWrite.cs
@@ -1,0 +1,23 @@
+﻿#if !PREVIOUS
+
+using BenchmarkDotNet.Attributes;
+using Snappier.Internal;
+
+namespace Snappier.Benchmarks
+{
+    public class VarIntEncodingWrite
+    {
+        [Params(0u, 256u, 65536u)]
+        public uint Value { get; set; }
+
+        readonly byte[] _dest = new byte[8];
+
+        [Benchmark(Baseline = true)]
+        public int Baseline()
+        {
+            return VarIntEncoding.Write(_dest, Value);
+        }
+    }
+}
+
+#endif