--- a/Snappier/Internal/SnappyCompressor.cs
+++ b/Snappier/Internal/SnappyCompressor.cs
@@ -18,7 +18,7 @@ namespace Snappier.Internal
 
             _workingMemory.EnsureCapacity(input.Length);
 
-            int bytesWritten = WriteUncompressedLength(output, input.Length);
+            int bytesWritten = VarIntEncoding.Write(output, (uint)input.Length);
             output = output.Slice(bytesWritten);
 
             while (input.Length > 0)
@@ -67,50 +67,6 @@ namespace Snappier.Internal
         {
             _workingMemory?.Dispose();
             _workingMemory = null;
-        }
-
-        private static int WriteUncompressedLength(Span<byte> output, int length)
-        {
-            const int b = 0b1000_0000;
-
-            unchecked
-            {
-                if (length < (1 << 7))
-                {
-                    output[0] = (byte) length;
-                    return 1;
-                }
-                else if (length < (1 << 14))
-                {
-                    output[0] = (byte) (length | b);
-                    output[1] = (byte) (length >> 7);
-                    return 2;
-                }
-                else if (length < (1 << 21))
-                {
-                    output[0] = (byte) (length | b);
-                    output[1] = (byte) ((length >> 7) | b);
-                    output[2] = (byte) (length >> 14);
-                    return 3;
-                }
-                else if (length < (1 << 28))
-                {
-                    output[0] = (byte) (length | b);
-                    output[1] = (byte) ((length >> 7) | b);
-                    output[2] = (byte) ((length >> 14) | b);
-                    output[3] = (byte) (length >> 21);
-                    return 4;
-                }
-                else
-                {
-                    output[0] = (byte) (length | b);
-                    output[1] = (byte) ((length >> 7) | b);
-                    output[2] = (byte) ((length >> 14) | b);
-                    output[3] = (byte) ((length >> 21) | b);
-                    output[4] = (byte) (length >> 28);
-                    return 5;
-                }
-            }
         }
 
         #region CompressFragment